+++ PSFramework/changelog.md
@@ -3,6 +3,7 @@
 ## X.X.X.X : 2018-05-27
  - New: Add command Remove-PSFNull, will clean the pipeline from unwanted empty objects.
  - New: Add command Test-PSFShouldProcess, implementing the `-Confirm` and `-WhatIf` parameters for a command. Useful to mock the test and make it more readable.
+ - New: Add command Test-PSFPowerShell, allowing simple powershell environment validation and mocking.
 
 ## 0.9.18.52 : 2018-05-20
  - New: Add command Import-PSFCmdlet, will register a cmdlet in PowerShell